VPLAY-12834 ParseSegmentIndexBox incorrect index
