+++ AampMPDUtils.cpp
@@ -18,8 +18,8 @@
 */
 
 #include "AampMPDUtils.h"
+#include "AampBoxReader.h"
 #include <inttypes.h>
-#include <type_traits>
 
 /**
  * @brief Get xml node form reader
@@ -211,76 +211,6 @@ double ComputeFragmentDuration( uint32_t duration, uint32_t timeScale )
 	}
 	return newduration;
 }
-
-namespace
-{
-/**
- * @brief Lightweight cursor for reading big-endian ISOBMFF box fields.
- *
- * Provides typed Read<T>() and Skip<T>() operations.
- * The caller is responsible for validating the box size field before
- * reading individual fields.
- */
-class BoxReader final
-{
-public:
-	constexpr explicit BoxReader(const uint8_t *data) noexcept
-		: mCursor{data}
-	{
-	}
-
-	/**
-	 * @brief Read a big-endian integer field and advance the cursor.
-	 * @tparam T  Integral type whose sizeof determines the field width.
-	 * @return    The value read in host byte order.
-	 */
-	template <typename T>
-	[[nodiscard]] T Read() noexcept
-	{
-		static_assert(std::is_integral_v<T>,
-					  "Read<T> requires an integral type");
-		static_assert(sizeof(T) == 1 || sizeof(T) == 2 ||
-					  sizeof(T) == 4 || sizeof(T) == 8,
-					  "Read<T> only supports 1/2/4/8-byte types");
-
-		uint64_t val{0};
-		for (size_t i = 0; i < sizeof(T); ++i)
-		{
-			val = (val << 8) | static_cast<uint8_t>(mCursor[i]);
-		}
-		mCursor += sizeof(T);
-		return static_cast<T>(val);
-	}
-
-	/**
-	 * @brief Skip a field without reading it.
-	 * @tparam T  Integral type whose sizeof determines the skip width.
-	 */
-	template <typename T>
-	void Skip() noexcept
-	{
-		static_assert(std::is_integral_v<T>,
-					  "Skip<T> requires an integral type");
-		static_assert(sizeof(T) == 1 || sizeof(T) == 2 ||
-					  sizeof(T) == 4 || sizeof(T) == 8,
-					  "Skip<T> only supports 1/2/4/8-byte types");
-
-		mCursor += sizeof(T);
-	}
-
-	/**
-	 * @brief Skip a runtime number of bytes.
-	 * @param n  Number of bytes to advance.
-	 */
-	void Skip(size_t n) noexcept
-	{
-		mCursor += n;
-	}
-
-private:
-	const uint8_t *mCursor;
-};
-} // namespace
 
 /**
  * @brief Parse segment index box
